remove utils file
